% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.expand_loc_error}
\alias{.expand_loc_error}
\title{Expand location error to match number of locations}
\usage{
.expand_loc_error(location_error, n_locs)
}
\description{
Expand location error to match number of locations
}
\keyword{internal}
