% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_move2}
\alias{.validate_move2}
\title{Validate a move2 object (common checks that apply regardless of track count)}
\usage{
.validate_move2(x)
}
\description{
Validate a move2 object (common checks that apply regardless of track count)
}
\keyword{internal}
